Support beautification of HTTP 100 responses.
Detect when the server sends an HTTP 100 (Continue) response header
and kick our beautification loop into 'continuing' mode.  While in
this mode, we wait for another ^HTTP/ header to be received before
responding to the `\r\n` break sequence signalling the end of the
headers and the start of the body.  This gives us output like:

    HTTP/1.1 100 (Continue)

    HTTP/1.1 200 OK
    Content-Length: 387
    Content-Type: application/json; charset=UTF-8

    {
    }

This comes into play when uploading files using curlish.

Previously, these types of HTTP conversations were not "beautified"
because the header processing loop bailed out after the initial
`HTTP/1.1 100 (Continue)\r\n\r\n` content was handled.